requirements.txt
python-dotenv==1.0.1
disnake==2.9.2 